Stop AOTC resources from forcibly showing up on the readout
Removes the force show for AOTC resources and Call of the Cosmos components. Having the AOTC resources and especially Call of the Cosmos components permanently on your resource readout when you are still in the tribal era is a bit jarring. This change aims to more seamlessly integrate AOTC into regular playthroughs.